+++ packages/eslint-plugin/configs/recommended-with-formatting.js
@@ -7,6 +7,7 @@ module.exports = {
 		require.resolve( './esnext.js' ),
 		require.resolve( './i18n.js' ),
 	],
+	plugins: [ 'import' ],
 	env: {
 		node: true,
 	},
@@ -15,16 +16,44 @@ module.exports = {
 		document: true,
 		wp: 'readonly',
 	},
+	rules: {
+		'import/no-extraneous-dependencies': [
+			'error',
+			{
+				devDependencies: false,
+				peerDependencies: true,
+			},
+		],
+		'import/no-unresolved': 'error',
+	},
 	overrides: [
 		{
 			// Unit test files and their helpers only.
 			files: [ '**/@(test|__tests__)/**/*.js', '**/?(*.)test.js' ],
 			extends: [ require.resolve( './test-unit.js' ) ],
+			rules: {
+				'import/no-extraneous-dependencies': [
+					'error',
+					{
+						devDependencies: true,
+						peerDependencies: true,
+					},
+				],
+			},
 		},
 		{
 			// End-to-end test files and their helpers only.
 			files: [ '**/specs/**/*.js', '**/?(*.)spec.js' ],
 			extends: [ require.resolve( './test-e2e.js' ) ],
+			rules: {
+				'import/no-extraneous-dependencies': [
+					'error',
+					{
+						devDependencies: true,
+						peerDependencies: true,
+					},
+				],
+			},
 		},
 	],
 };